+++ controllers/o1/include/Motion.hpp
@@ -8,16 +8,16 @@
 // PID Controller
 class PID {
    private:
-    float min_out, max_out, min_integral, max_integral;
-    float output_speed;
-    float proportional, integral, derivative, prev_error;
-    float kp, ki, kd;
-    const float time_pid = 0.01;
+    double min_out, max_out, min_integral, max_integral;
+    double output_speed;
+    double proportional, integral, derivative, prev_error;
+    double kp, ki, kd;
+    const double time_pid = 0.01;
 
    public:
     PID();
-    float calculatePID(float error, float min_max);
-    void setParam(float kp_, float ki_, float kd_);
+    double calculatePID(double error, double min_max);
+    void setParam(double kp_, double ki_, double kd_);
     void reset();
 };
 
@@ -26,7 +26,7 @@ class Motion {
    private:
     std::vector<double> output = {0, 0, 0, 0};
     std::vector<double> error = {0, 0, 0, 0};
-    void basicMotion(float vx, float vy, float vtheta, float thetaRobot,
+    void basicMotion(double vx, double vy, double vtheta, double thetaRobot,
                      Point2D &output);
     void accelControl(Point2D *out, Point2D &data);
 
@@ -35,10 +35,11 @@ class Motion {
     Motion();
     void positionAngularControl(double &errorX, double &errorY,
                                 double &errorTheta, double yaw,
-                                Point2D &outMotor,  double nearestX, 
+                                Point2D &outMotor, double nearestX,
                                 double nearestY, double minDistance,
-                                double currentX, double currentY, Point2D targetPos,
-                                int count, std::vector<Point2D> &path);
+                                double currentX, double currentY,
+                                Point2D targetPos, int count,
+                                std::vector<Point2D> &path);
 };
 
 #endif
