+++ controllers/o1/include/Robot.hpp
@@ -14,11 +14,11 @@ class RobotKinematic {
     Point2D vel = {0, 0, 0};
 
     static RobotKinematic *instance;
-    float a = 45;  // Angle between wheels
-    float L =
+    double a = 45;  // Angle between wheels
+    double L =
         41.6;  // radius from center to wheel (cm); BaseLength 4,795.94 MM;
-    float r_wheel = 6;  // wheel radius (cm)
-    float circumference = 2 * M_PI * r_wheel;
+    double r_wheel = 6;  // wheel radius (cm)
+    double circumference = 2 * M_PI * r_wheel;
     std::vector<double> Venc = {0, 0, 0, 0};
     std::vector<double> prevEnc = {0, 0, 0, 0};
 
@@ -28,14 +28,14 @@ class RobotKinematic {
 
     RobotKinematic(){};
     Point2D getPos();
-    void forwardKinematics(Point2D &outForward, float s1, float s2, float s3,
-                           float s4);
+    void forwardKinematics(Point2D &outForward, double s1, double s2, double s3,
+                           double s4);
     void calculateOdometry(const double yaw);
-    void inverseKinematics(wheelAngularVel &outputInverse, float vx, float vy,
-                           float theta);
+    void inverseKinematics(wheelAngularVel &outputInverse, double vx, double vy,
+                           double theta);
     double angleNormalize(double angle);
 
-    void setInitialPosition(float x, float y, float theta);
+    void setInitialPosition(double x, double y, double theta);
 };
 };  // namespace robot
