+++ controllers/o1/include/Visualize.hpp
@@ -10,12 +10,12 @@
 class Visualize {
    private:
     // Field Dimensions
-    const float A = 22, B = 14, C = 6.9, D = 3.9, E = 2.25, F = .75, G = .75,
-                H = 4, I = 3.6, J = .15, K = .125, L = 1, M = 1, N = 8, O = 1,
-                P = .5, Q = 3.5, r = 0.5,
-                s = 2.4;  // r = width of goalpost, s = length of goalpost
-    float resX, resY;     // Image Resolution
-    float scale;          // Scaling Factor
+    const double A = 22, B = 14, C = 6.9, D = 3.9, E = 2.25, F = .75, G = .75,
+                 H = 4, I = 3.6, J = .15, K = .125, L = 1, M = 1, N = 8, O = 1,
+                 P = .5, Q = 3.5, r = 0.5,
+                 s = 2.4;  // r = width of goalpost, s = length of goalpost
+    double resX, resY;     // Image Resolution
+    double scale;          // Scaling Factor
 
     //  void drawArrow(cv::Mat &image, const cv::Point &pStart,
     //                 const cv::Point &pEnd, const cv::Scalar &color,
@@ -28,7 +28,7 @@ class Visualize {
                       double tipLength);
 
    public:
-    Visualize(float resolutionX);
+    Visualize(double resolutionX);
     void visualizeGame(std::vector<Point2D> &path, Point2D &nowPos, int count,
                        double yaw, std::vector<Point2D> &obstacles,
                        Point2D &ball);
